feat(todos): send messages to todos by reacting

+++ src/events/todos/reaction.event.js
@@ -1,3 +1,4 @@
+const { TextChannel } = require('discord.js');
 const { EventHandler } = require('sensum');
 
 module.exports = new EventHandler({
@@ -8,12 +9,16 @@ module.exports = new EventHandler({
         reaction.message.delete().catch(console.log);
       }
     } else {
-      const msg = message.content.replace('todo:', '').trim();
-      const c = bot.guilds.cache.get('683463960501944327').channels.cache.get('888472757900288090');
-      if (c instanceof TextChannel) {
-        c.send(`- ${msg}`)
-          .then(() => reaction.message.delete())
-          .catch(() => {});
+      if (['☑️', '✔️', '✅', '❤️'].includes(reaction.emoji.name)) {
+        const msg = reaction.message.content.replace('todo:', '').trim();
+        const c = bot.guilds.cache
+          .get('683463960501944327')
+          .channels.cache.get('888472757900288090');
+        if (c instanceof TextChannel) {
+          c.send(`- ${msg}`)
+            .then(() => reaction.message.delete())
+            .catch(console.log);
+        }
       }
     }
   },
